Add export flags
