+++ Sources/Queues/Exports.swift
@@ -1,3 +1,5 @@
+#if !BUILDING_DOCC
+
 @_exported import struct Foundation.Date
 @_exported import struct Logging.Logger
 @_exported import class NIO.EventLoopFuture
@@ -5,16 +7,4 @@
 @_exported import protocol NIO.EventLoop
 @_exported import struct NIO.TimeAmount
 
-import class NIO.RepeatedTask
-
-extension RepeatedTask {
-    func syncCancel(on eventLoop: EventLoop) {
-        do {
-            let promise = eventLoop.makePromise(of: Void.self)
-            self.cancel(promise: promise)
-            try promise.futureResult.wait()
-        } catch {
-            print("failed cancelling repeated task \(error)")
-        }
-    }
-}
+#endif
